Will be moving code to GitHub soon
Updated to Spearmint 378

- Spearmint 376

Restore botlist command, I removed it by mistake.

- Spearmint 377

ioquake3 resync to revision 2395 from 2392.

- Fix spectator client being switched from follow to free after map_restart if following a client with a higher client number.
- Fix client unlinking issue caused by ent->s.number being set to followed client's ps->clientNum after map_restart. Reported by Ensiform.
- In G_AddBot, try to allocate clientNum before doing anything else.
- In G_AddBot, don't set SVF_BOT and inuse. It's done in ClientConnect, plus inuse causes ClientDisconnect to be run for no reason.
- In G_AddBot, only set skill in bot useinfo once.
- Avoid using cl->ps.clientNum to check if cl is a bot.
- Fix bot skill format so it doesn't always have a space at the beginning of it.

- Spearmint 378

ioquake3 resync to revision 2398 from 2395.

More fixes to the macosx buildsystem.  This removes the SDL Framework and makes use of a SDL library that is position independant.  This also brings back PPC builds into the UB and also as a standa alone build choice.
Have make-macosx.sh require the user to specify which architecture she/he wants to build for and suggest building UB's if the user is unaware of what architectures are
Lets list all the valid options.